+++ src/routes/currency/[id]/+page.ts
@@ -3,13 +3,6 @@ import type { PageLoad } from './$types';
 export const load: PageLoad = async ({ parent, params }) => {
 	const { supabase } = await parent();
 
-	const { data: currency, error: currencyError } = await supabase
-		.from('currencies')
-		.select()
-		.eq('id', params.id)
-		.single();
-	if (currencyError) console.log(currencyError.message);
-
 	const { data: venues, error: venuesError } = await supabase
 		.from('venues')
 		.select()
@@ -26,7 +19,6 @@ export const load: PageLoad = async ({ parent, params }) => {
 	const count = tokens?.length ?? null;
 
 	return {
-		currency: currency,
 		venues: venues,
 		count: count
 	};