+++ src/routes/currency/[id]/transactions/+page.svelte
@@ -4,15 +4,15 @@
 	import Transactions from '$lib/components/Transactions.svelte';
 
 	let { data } = $props();
-	let { venue, transactions } = $derived(data);
+	let { currency, transactions } = $derived(data);
 </script>
 
-{#if venue && transactions}
-	<Page title={venue.title} subtitle="Transactions">
-		<p>These are all {transactions.length} transactions for this venue.</p>
+{#if currency && transactions}
+	<Page title={currency.name} subtitle="Transactions">
+		<p>These are all {transactions.length} transactions for this currency.</p>
 		<Transactions {transactions} />
 	</Page>
-{:else if venue === null}
+{:else if currency === null}
 	<h1>Oops.</h1>
 	<Feedback error>Unknown scholar.</Feedback>
 {:else if transactions === null}